requirements.txt
flask-socketio
flask-mysql